refactor: modularize large components — extract 4 hooks + hmTime util
- useNowTick: aprimorado com alinhamento de tick; ClientCard.tsx migrado para import compartilhado
- utils/hmTime: extrai pad2/toMinutes/fromMinutes/weekdayLabel do QuickScheduleModal
- useQuickScheduleSave: extrai handleSave (~270 linhas) de QuickScheduleModal; gerencia saving/error state
- useAgendaModals: consolida 11 estados de modal + openMonthly/openWeekly/openDaily + event bridges (openScheduleEdit, openDailyAgenda, openAppointmentDetails)
- usePendingActionsListeners: consolida 5 useEffects de pending actions (confirmFinalize, openPending, forceClose, auto-close, debug) de Home.tsx
- Home.tsx: reduzido de ~1070 para ~300 linhas de lógica efetiva; route params, postLoginFocus e systemMessage preservados
- QuickScheduleModal.tsx: reduzido de ~1103 para ~750 linhas
- Zero erros TypeScript em todos os 8 arquivos modificados
