+++ frontend/src/pages/Home.tsx
@@ -1,7 +1,5 @@
 // src/pages/Home.tsx
 import React, { useEffect, useState } from 'react';
-import { on } from '../events/bus';
-import type { OpenDailyAgendaPayload } from '../events/bus';
 
 import Header from '../components/Header';
 import Faixa from '../components/Faixa';
@@ -18,42 +16,45 @@ import SystemMessageModal from '../components/SystemMessageModal';
 import DailyAgendaModal from '../components/DailyAgendaModal';
 import AppointmentDetailsModal from '../components/AppointmentDetailsModal';
 import PendingActionsModal from '../components/PendingActionsModal';
-import { getAppointmentOverride } from '../utils/appointments/overrides';
-import type { SharedAppointmentLike } from '../components/shared/AppointmentCard';
-import type { ClientBasic } from '../types/ClientBasic';
-import { API_BASE } from '../config/api';
-import { isTokenExpired } from '../utils/jwt';
 import type { Appointment } from '../hooks/useAppointments';
 import { useAppVersionWatcher, acceptAndReload } from '../hooks/useAppVersion';
 import { useAppointmentsLivePing } from '../hooks/useAppointmentsLivePing';
 import { dispatchers } from '../events/dispatchers';
 import { focusClientCard } from '../utils/focusClientCard';
+import { useAgendaModals, ensureClientBasic } from '../hooks/useAgendaModals';
+import { API_BASE } from '../config/api';
+import { usePendingActionsListeners } from '../hooks/usePendingActionsListeners';
 
 export default function Home() {
     const [selectedClientId, setSelectedClientId] = useState<number | null>(
         null,
     );
-    // Route-driven agenda modal states
-    // const [scheduleOpen, setScheduleOpen] = useState(false); // removido
-    const [monthlyOpen, setMonthlyOpen] = useState(false);
-    const [weeklyOpen, setWeeklyOpen] = useState(false);
-    const [routeClient, setRouteClient] = useState<ClientBasic | undefined>(
-        undefined,
-    );
-    // routeDefaultDate removido com ScheduleModal
-    const [routeEditAppt, setRouteEditAppt] = useState<Appointment | null>(
-        null,
-    );
-    const [dailyOpen, setDailyOpen] = useState(false);
-    const [dailyDate, setDailyDate] = useState<Date>(new Date());
-    const [dailyFocusId, setDailyFocusId] = useState<number | undefined>(
-        undefined,
-    );
-    const [detailsOpen, setDetailsOpen] = useState(false);
-    const [detailsAppt, setDetailsAppt] = useState<Appointment | null>(null);
-    const [routeInitialMonth, setRouteInitialMonth] = useState<
-        Date | undefined
-    >(undefined);
+    const {
+        monthlyOpen,
+        setMonthlyOpen,
+        routeClient,
+        setRouteClient,
+        routeInitialMonth,
+        setRouteInitialMonth,
+        weeklyOpen,
+        setWeeklyOpen,
+        quickOpen,
+        setQuickOpen,
+        routeEditAppt,
+        setRouteEditAppt,
+        dailyOpen,
+        setDailyOpen,
+        dailyDate,
+        dailyFocusId,
+        detailsOpen,
+        setDetailsOpen,
+        detailsAppt,
+        setDetailsAppt,
+        openMonthly,
+        openWeekly,
+        openDaily,
+        clearAgendaRouteFlags,
+    } = useAgendaModals();
     const [sysMsg, setSysMsg] = useState<{
         text: string;
         type: 'success' | 'error' | 'info';
@@ -63,80 +64,8 @@ export default function Home() {
     // Live ping: simple heuristic — enable while the page is open.
     // We can refine to enable only when there may be ongoing appointments by inspecting clients in future.
     useAppointmentsLivePing({ enabled: true, pollIntervalMs: 30000 });
-    // Pending actions modal state (replaces ConfirmFinalizeModal flow)
-    const [pendingActionsOpen, setPendingActionsOpen] = useState(false);
-    const [pendingAppt, setPendingAppt] =
-        useState<SharedAppointmentLike | null>(null);
-
-    // (reminderModal removido — push notification foca o ClientCard diretamente)
-    // Após fechamento, limpa o objeto pendente pouco depois para evitar manter props antigas vivas
-    useEffect(() => {
-        if (pendingActionsOpen) return;
-        const t = setTimeout(() => {
-            setPendingAppt(pa => (pendingActionsOpen ? pa : null));
-        }, 80);
-        return () => clearTimeout(t);
-    }, [pendingActionsOpen]);
-    // Suppress rapid reopen of PendingActions for a short window after closing
-    const lastPendingCloseRef = React.useRef<number>(0);
-
-    // Aux: carrega nome básico do cliente dado um ID (cache localStorage se possível)
-    async function ensureClientBasic(id: number): Promise<ClientBasic> {
-        const cached = localStorage.getItem(`client.name.${id}`);
-        if (cached) {
-            const [first_name, ...rest] = cached.split(' ');
-            const last_name = rest.join(' ');
-            return {
-                id,
-                first_name,
-                last_name,
-                phone: '',
-                email: '',
-            } as ClientBasic;
-        }
-        const token = localStorage.getItem('accessToken');
-        if (isTokenExpired(token))
-            return {
-                id,
-                first_name: 'Cliente',
-                last_name: String(id),
-                phone: '',
-                email: '',
-            } as ClientBasic;
-        try {
-            const res = await fetch(`${API_BASE}/register/clients/${id}/`, {
-                headers: { Authorization: `Bearer ${token}` },
-            });
-            if (res.ok) {
-                const data = await res.json();
-                const cb: ClientBasic = {
-                    id: data.id,
-                    first_name: data.first_name || 'Cliente',
-                    last_name: data.last_name || '',
-                    phone: data.phone || '',
-                    email: data.email || '',
-                };
-                try {
-                    localStorage.setItem(
-                        `client.name.${id}`,
-                        `${cb.first_name} ${cb.last_name}`.trim(),
-                    );
-                } catch {
-                    /* noop */
-                }
-                return cb;
-            }
-        } catch {
-            /* noop */
-        }
-        return {
-            id,
-            first_name: 'Cliente',
-            last_name: String(id),
-            phone: '',
-            email: '',
-        } as ClientBasic;
-    }
+    const { pendingActionsOpen, pendingAppt, closePendingActions } =
+        usePendingActionsListeners();
 
     // Seleciona o cliente via ?client=ID e abre modais conforme params (?new, ?edit, ?mode)
     useEffect(() => {
@@ -156,10 +85,8 @@ export default function Home() {
                     ? new Date(dateStr + 'T00:00:00')
                     : undefined;
                 if (parsedDate && !Number.isNaN(parsedDate.getTime())) {
-                    // setRouteDefaultDate(parsedDate);
                     setRouteInitialMonth(parsedDate);
                 } else {
-                    // setRouteDefaultDate(undefined);
                     setRouteInitialMonth(undefined);
                 }
 
@@ -170,8 +97,10 @@ export default function Home() {
                 // Push notification reminder: focus ClientCard (or save pending for post-login)
                 if (reminderId) {
                     const token = localStorage.getItem('accessToken');
+                    const isWa = url.searchParams.get('wa') === '1';
+                    const waPhone = url.searchParams.get('wp') ?? '';
+                    const waText = url.searchParams.get('wt') ?? '';
                     if (token) {
-                        // Logged in → resolve client_id and focus card
                         try {
                             const res = await fetch(
                                 `${API_BASE}/agenda/appointments/${reminderId}/`,
@@ -192,33 +121,51 @@ export default function Home() {
                                     setSelectedClientId(clientId);
                                     focusClientCard(clientId, { delayMs: 300 });
                                 }
+                                if (isWa) {
+                                    fetch(
+                                        `${API_BASE}/agenda/appointments/${reminderId}/confirm-whatsapp/`,
+                                        {
+                                            method: 'POST',
+                                            headers: {
+                                                Authorization: `Bearer ${token}`,
+                                                'Content-Type':
+                                                    'application/json',
+                                            },
+                                        },
+                                    ).catch(() => {
+                                        /* silencioso */
+                                    });
+                                    if (waPhone && waText) {
+                                        window.open(
+                                            `https://wa.me/${waPhone}?text=${encodeURIComponent(waText)}`,
+                                            '_blank',
+                                        );
+                                    }
+                                }
                             }
                         } catch {
                             /* silencioso */
                         }
                     } else {
-                        // Not logged in → save to sessionStorage; useEffect pós-login will pick it up
-                        const waPhone = url.searchParams.get('wp') ?? '';
-                        const waText = url.searchParams.get('wt') ?? '';
                         sessionStorage.setItem(
                             'pushPending',
                             JSON.stringify({
                                 appointmentId: Number(reminderId),
+                                wa: isWa,
                                 waPhone,
                                 waText,
                             }),
                         );
                     }
-                    // Clean up URL params without reloading
                     const clean = new URL(window.location.href);
                     clean.searchParams.delete('reminder');
+                    clean.searchParams.delete('wa');
                     clean.searchParams.delete('wp');
                     clean.searchParams.delete('wt');
                     window.history.replaceState({}, '', clean.toString());
                 }
 
                 if (editId && cid) {
-                    // Carrega cliente e compromisso para abrir direto em edição (QuickSchedule)
                     const clientBasic = await ensureClientBasic(Number(cid));
                     setRouteClient(clientBasic);
                     try {
@@ -233,9 +180,8 @@ export default function Home() {
                                     },
                                 },
                             );
-                            if (res.ok) {
+                            if (res.ok)
                                 appt = (await res.json()) as Appointment;
-                            }
                         }
                         if (appt) setRouteEditAppt(appt);
                     } catch {
@@ -258,7 +204,7 @@ export default function Home() {
                     setMonthlyOpen(true);
                 }
             } catch {
-                // ignore
+                /* ignore */
             }
         })();
     }, []);
@@ -273,6 +219,7 @@ export default function Home() {
             if (!token) return;
             let pending: {
                 appointmentId: number;
+                wa?: boolean;
                 waPhone: string;
                 waText: string;
             };
@@ -299,6 +246,26 @@ export default function Home() {
                         if (clientId) {
                             setSelectedClientId(clientId);
                             focusClientCard(clientId, { delayMs: 400 });
+                        }
+                        if (pending.wa) {
+                            fetch(
+                                `${API_BASE}/agenda/appointments/${pending.appointmentId}/confirm-whatsapp/`,
+                                {
+                                    method: 'POST',
+                                    headers: {
+                                        Authorization: `Bearer ${token}`,
+                                        'Content-Type': 'application/json',
+                                    },
+                                },
+                            ).catch(() => {
+                                /* silencioso */
+                            });
+                            if (pending.waPhone && pending.waText) {
+                                window.open(
+                                    `https://wa.me/${pending.waPhone}?text=${encodeURIComponent(pending.waText)}`,
+                                    '_blank',
+                                );
+                            }
                         }
                     },
                 )
@@ -360,373 +327,9 @@ export default function Home() {
         }
     }, []);
 
-    // Listener global: ao clicar em "Finalizar", abre o PendingActionsModal
-    useEffect(() => {
-        async function onConfirmFinalize(e: Event) {
-            const ce = e as CustomEvent;
-            const det =
-                (ce && (ce as CustomEvent).detail) ||
-                ({} as {
-                    isEarly?: boolean;
-                    clientId?: number;
-                    appointmentId?: number | null;
-                    proceed?: () => void; // legacy, não usamos mais
-                });
-            // Bloqueia o default para impedir fallback com window.confirm
-            try {
-                (e as Event).preventDefault?.();
-            } catch {
-                /* noop */
-            }
-            const apptId = det.appointmentId;
-            if (!apptId) {
-                try {
-                    window.dispatchEvent(
-                        new CustomEvent('systemMessage', {
-                            detail: {
-                                text: 'Não foi possível identificar o agendamento para concluir.',
-                                type: 'error',
-                            },
-                        }),
-                    );
-                } catch {
-                    /* noop */
-                }
-                return;
-            }
-            // Debounce/suppress: if we just closed pendingActions, ignore a rapid duplicate open
-            const nowTs = Date.now();
-            if (nowTs - lastPendingCloseRef.current < 1500) {
-                try {
-                    window.dispatchEvent(
-                        new CustomEvent('debug:log', {
-                            detail: {
-                                label: 'Home: suppressed confirmFinalize due to recent forceClose',
-                                data: {
-                                    deltaMs:
-                                        nowTs - lastPendingCloseRef.current,
-                                },
-                                ts: nowTs,
-                            },
-                        }),
-                    );
-                } catch {
-                    /* noop */
-                }
-                return;
-            }
-            // Carrega dados do agendamento para exibir no modal de Ações Pendentes
-            try {
-                const token = localStorage.getItem('accessToken') || '';
-                const headers: Record<string, string> = {};
-                if (token) headers['Authorization'] = `Bearer ${token}`;
-                const r = await fetch(
-                    `${API_BASE}/agenda/appointments/${apptId}/`,
-                    { headers, cache: 'no-store' },
-                );
-                if (!r.ok) throw new Error('Falha ao carregar agendamento');
-                const data = await r.json();
-                const appt: SharedAppointmentLike = {
-                    id: data.id,
-                    start_at: data.start_at,
-                    end_at: data.end_at,
-                    status: data.status,
-                    notes: data.notes,
-                    client_name: data.client_name,
-                    client: data.client,
-                    title: data.title,
-                };
-                setPendingAppt(appt);
-                setPendingActionsOpen(true);
-            } catch (err) {
-                const msg =
-                    err instanceof Error
-                        ? err.message
-                        : 'Erro ao abrir ações pendentes';
-                try {
-                    window.dispatchEvent(
-                        new CustomEvent('systemMessage', {
-                            detail: { text: msg, type: 'error' },
-                        }),
-                    );
-                } catch {
-                    /* noop */
-                }
-            }
-        }
-        window.addEventListener(
-            'confirmFinalizeAppointment',
-            onConfirmFinalize as EventListener,
-        );
-        return () =>
-            window.removeEventListener(
-                'confirmFinalizeAppointment',
-                onConfirmFinalize as EventListener,
-            );
-    }, []);
+    // Aberturas diretas dos modais da Agenda
+    // openMonthly, openWeekly, openDaily e clearAgendaRouteFlags definidos em useAgendaModals
 
-    // Novo handler global: qualquer componente pode abrir Ações Pendentes via evento
-    useEffect(() => {
-        // Mantém supressões temporárias pós-cancel
-        const recentCanceled = new Map<number, number>(); // apptId -> timestamp
-        // Listener auxiliar para marcar cancelamentos concluídos
-        function onCancelFinalizeFlag(ev: Event) {
-            const ce = ev as CustomEvent;
-            const det = (ce?.detail || {}) as { id?: number; status?: string };
-            if (det?.id && det.status === 'canceled') {
-                recentCanceled.set(det.id, Date.now());
-            }
-        }
-        window.addEventListener(
-            'appointment:statusChanged',
-            onCancelFinalizeFlag,
-        );
-
-        async function onOpenPending(e: Event) {
-            const ce = e as CustomEvent;
-            const det =
-                (ce && (ce as CustomEvent).detail) ||
-                ({} as {
-                    appt?: SharedAppointmentLike;
-                    appointmentId?: number | null;
-                });
-            try {
-                const stack = new Error().stack;
-                window.dispatchEvent(
-                    new CustomEvent('debug:log', {
-                        detail: {
-                            label: 'Home: pendingActions:open received',
-                            data: {
-                                hasApptInDetail: !!det.appt,
-                                appointmentId:
-                                    det.appt?.id ?? det.appointmentId,
-                                triggeredAt: new Date().toISOString(),
-                                stack,
-                            },
-                            ts: Date.now(),
-                        },
-                    }),
-                );
-            } catch {
-                /* noop */
-            }
-            try {
-                (e as Event).preventDefault?.();
-            } catch {
-                /* noop */
-            }
-            const nowTs = Date.now();
-            if (nowTs - lastPendingCloseRef.current < 1500) {
-                try {
-                    window.dispatchEvent(
-                        new CustomEvent('debug:log', {
-                            detail: {
-                                label: 'Home: suppressed pendingActions:open due to recent forceClose',
-                                data: {
-                                    deltaMs:
-                                        nowTs - lastPendingCloseRef.current,
-                                },
-                                ts: nowTs,
-                            },
-                        }),
-                    );
-                } catch {
-                    /* noop */
-                }
-                return;
-            }
-            if (det.appt) {
-                // Se acabou de ser cancelado, suprimir reabertura
-                const rcTs = det.appt.id
-                    ? recentCanceled.get(det.appt.id)
-                    : undefined;
-                if (rcTs && Date.now() - rcTs < 8000) {
-                    try {
-                        window.dispatchEvent(
-                            new CustomEvent('debug:log', {
-                                detail: {
-                                    label: 'Home: suppressed reopen (recent canceled appt object)',
-                                    data: {
-                                        apptId: det.appt.id,
-                                        deltaMs: Date.now() - rcTs,
-                                    },
-                                    ts: Date.now(),
-                                },
-                            }),
-                        );
-                    } catch {
-                        /* noop */
-                    }
-                    return;
-                }
-                setPendingAppt(det.appt);
-                setPendingActionsOpen(true);
-                return;
-            }
-            const apptId = det.appointmentId;
-            if (!apptId) return;
-            const rcTs = recentCanceled.get(apptId);
-            if (rcTs && Date.now() - rcTs < 8000) {
-                try {
-                    window.dispatchEvent(
-                        new CustomEvent('debug:log', {
-                            detail: {
-                                label: 'Home: suppressed reopen (recent canceled apptId)',
-                                data: { apptId, deltaMs: Date.now() - rcTs },
-                                ts: Date.now(),
-                            },
-                        }),
-                    );
-                } catch {
-                    /* noop */
-                }
-                return;
-            }
-            try {
-                const token = localStorage.getItem('accessToken') || '';
-                const headers: Record<string, string> = {};
-                if (token) headers['Authorization'] = `Bearer ${token}`;
-                const r = await fetch(
-                    `${API_BASE}/agenda/appointments/${apptId}/`,
-                    { headers, cache: 'no-store' },
-                );
-                if (!r.ok) throw new Error('Falha ao carregar agendamento');
-                const data = await r.json();
-                if (data.status && data.status !== 'scheduled') {
-                    // Não abrir para status cancelado/concluído
-                    try {
-                        window.dispatchEvent(
-                            new CustomEvent('debug:log', {
-                                detail: {
-                                    label: 'Home: skipped open (status not scheduled)',
-                                    data: { apptId, status: data.status },
-                                    ts: Date.now(),
-                                },
-                            }),
-                        );
-                    } catch {
-                        /* noop */
-                    }
-                    return;
-                }
-                const appt: SharedAppointmentLike = {
-                    id: data.id,
-                    start_at: data.start_at,
-                    end_at: data.end_at,
-                    status: data.status,
-                    notes: data.notes,
-                    client_name: data.client_name,
-                    client: data.client,
-                    title: data.title,
-                };
-                setPendingAppt(appt);
-                setPendingActionsOpen(true);
-            } catch (err) {
-                const msg =
-                    err instanceof Error
-                        ? err.message
-                        : 'Erro ao abrir ações pendentes';
-                try {
-                    window.dispatchEvent(
-                        new CustomEvent('systemMessage', {
-                            detail: { text: msg, type: 'error' },
-                        }),
-                    );
-                } catch {
-                    /* noop */
-                }
-            }
-        }
-        window.addEventListener(
-            'pendingActions:open',
-            onOpenPending as EventListener,
-        );
-        return () => {
-            window.removeEventListener(
-                'pendingActions:open',
-                onOpenPending as EventListener,
-            );
-            window.removeEventListener(
-                'appointment:statusChanged',
-                onCancelFinalizeFlag as EventListener,
-            );
-        };
-    }, []);
-
-    // Failsafe: se algum modal PendingActions disparar forceClose, alinhe o estado local
-    useEffect(() => {
-        const onForceClose = () => {
-            setPendingActionsOpen(false);
-            lastPendingCloseRef.current = Date.now();
-        };
-        window.addEventListener('pendingActions:forceClose', onForceClose);
-        return () =>
-            window.removeEventListener(
-                'pendingActions:forceClose',
-                onForceClose,
-            );
-    }, []);
-    // Auto-close if an override terminal status appears for the currently open pending appointment
-    useEffect(() => {
-        if (!pendingActionsOpen || !pendingAppt) return;
-        try {
-            const ov = getAppointmentOverride(pendingAppt.id as number);
-            if (ov && (ov.status === 'done' || ov.status === 'canceled')) {
-                setPendingActionsOpen(false);
-            }
-        } catch {
-            /* noop */
-        }
-    }, [pendingActionsOpen, pendingAppt]);
-
-    // Debug instrumentation: expose helper functions to window for diagnosing stuck PendingActionsModal
-    useEffect(() => {
-        const w = window as unknown as Record<string, unknown>;
-        w.__closePendingActions = () => {
-            try {
-                setPendingActionsOpen(false);
-                console.debug('[PendingDebug] __closePendingActions called');
-            } catch (e) {
-                console.warn('[PendingDebug] close error', e);
-            }
-        };
-        w.__dumpPendingActions = () => {
-            try {
-                const id = pendingAppt?.id;
-                const ov = id ? getAppointmentOverride(id) : undefined;
-                console.debug('[PendingDebug] dump', {
-                    pendingActionsOpen,
-                    apptId: id,
-                    apptStatus: pendingAppt?.status,
-                    override: ov,
-                });
-                return {
-                    pendingActionsOpen,
-                    id,
-                    status: pendingAppt?.status,
-                    override: ov,
-                };
-            } catch (e) {
-                console.warn('[PendingDebug] dump error', e);
-                return null;
-            }
-        };
-        const onDebugDump = () => {
-            (w.__dumpPendingActions as () => unknown)();
-        };
-        window.addEventListener('pendingActions:debugDump', onDebugDump);
-        return () => {
-            try {
-                delete w.__closePendingActions;
-                delete w.__dumpPendingActions;
-            } catch {
-                /* noop */
-            }
-            window.removeEventListener('pendingActions:debugDump', onDebugDump);
-        };
-    }, [pendingActionsOpen, pendingAppt]);
-
-    // Mensagem ao voltar da Agenda sem cliente selecionado
     useEffect(() => {
         try {
             const msg = localStorage.getItem('agenda.promptSelectClient');
@@ -783,122 +386,9 @@ export default function Home() {
 
     // Aberturas diretas dos modais da Agenda (novo fluxo vindo do NavBar)
     // openSchedule removido — consolidado no QuickSchedule
+    // openMonthly, openWeekly, openDaily e clearAgendaRouteFlags definidos em useAgendaModals
 
-    const openMonthly = async (clientId: number, date?: Date) => {
-        // Evita abrir sem cliente válido
-        if (!clientId || clientId <= 0) {
-            try {
-                window.dispatchEvent(
-                    new CustomEvent('systemMessage', {
-                        detail: {
-                            text: 'Selecione um cliente para abrir a Agenda Mensal.',
-                            type: 'info',
-                            autoCloseMs: 6000,
-                        },
-                    }),
-                );
-            } catch {
-                /* noop */
-            }
-            return;
-        }
-        const c = await ensureClientBasic(clientId);
-        setRouteClient(c);
-        setRouteInitialMonth(date);
-        setRouteEditAppt(null);
-        setMonthlyOpen(true);
-        setWeeklyOpen(false);
-    };
-
-    const openWeekly = (_date?: Date) => {
-        void _date;
-        // WeeklyAgendaModal atual não usa data; reservado para evolução
-        setRouteEditAppt(null);
-        try {
-            console.debug('[Home] Opening WeeklyAgendaModal');
-        } catch {
-            /* noop */
-        }
-        setWeeklyOpen(true);
-        setMonthlyOpen(false);
-    };
-
-    // Abrir visão diária (sem exigir cliente específico)
-    function openDaily(date: Date, focusId?: number) {
-        setDailyDate(date);
-        setDailyFocusId(focusId);
-        setDailyOpen(true);
-    }
-
-    // Remove parâmetros que forçam reabertura de modais (/agenda?new=1&edit=ID&mode=week)
-    function clearAgendaRouteFlags() {
-        try {
-            const url = new URL(window.location.href);
-            ['new', 'edit', 'mode', 'date'].forEach(k =>
-                url.searchParams.delete(k),
-            );
-            const params = url.searchParams.toString();
-            window.history.replaceState(
-                {},
-                '',
-                url.pathname + (params ? `?${params}` : ''),
-            );
-        } catch {
-            /* noop */
-        }
-    }
-
-    // Event bridge: abrir edição direta do QuickSchedule a partir do Monthly
-    useEffect(() => {
-        function onOpenScheduleEdit(e: CustomEvent) {
-            const det = (e && (e as CustomEvent).detail) || {};
-            const c = det.client as ClientBasic | undefined;
-            const a = det.appointment as Appointment | undefined;
-            if (!c || !c.id) return;
-            setRouteClient(c);
-            setRouteEditAppt(a ?? null);
-            setQuickOpen(true);
-            setMonthlyOpen(false);
-            setWeeklyOpen(false);
-        }
-        window.addEventListener(
-            'openScheduleEdit',
-            onOpenScheduleEdit as EventListener,
-        );
-        const disposeDaily = on('openDailyAgenda', det => {
-            const ext = det as OpenDailyAgendaPayload & {
-                focusAppointmentId?: number;
-            };
-            const dateIso = ext && ext.date; // backward compat
-            const focusAppointmentId = ext.focusAppointmentId;
-            if (dateIso) {
-                const d = new Date(dateIso);
-                if (!isNaN(d.getTime())) openDaily(d, focusAppointmentId);
-            } else openDaily(new Date(), focusAppointmentId);
-        });
-        const disposeDetails = on('openAppointmentDetails', det => {
-            const appt: Appointment | undefined = (
-                det as { appointment?: Appointment }
-            ).appointment;
-            if (appt) {
-                setDetailsAppt(appt);
-                setDetailsOpen(true);
-            }
-        });
-        return () => {
-            window.removeEventListener(
-                'openScheduleEdit',
-                onOpenScheduleEdit as EventListener,
-            );
-            disposeDaily();
-            disposeDetails();
-        };
-    }, []);
-
-    // Estado do Quick
-    const [quickOpen, setQuickOpen] = useState(false);
-
-    // Safeguard: if any previous modal left the page locked, ensure unlock on mount
+    // Listener global para mensagens do sistema
     useEffect(() => {
         try {
             window.dispatchEvent(new Event('ensureScrollUnlocked'));
@@ -1013,9 +503,7 @@ export default function Home() {
                     <PendingActionsModal
                         open
                         appt={pendingAppt}
-                        onClose={() => {
-                            setPendingActionsOpen(false);
-                        }}
+                        onClose={closePendingActions}
                     />
                 )}
                 {/* Reminder: push notification click focuses the ClientCard directly (no modal) */}
